requirements.txt
rich
keyboard
requests
Pillow
